requirements.txt
py-cord==2.0.1
mongoengine==0.24.2
numpy==1.23.1
pandas==1.4.3
python-dotenv==0.20.0
aiosqlite==0.17.0

